Removed .py command
Removed .py command. allows you to hijack the bot and have a user say things into chat. There is no ACL for this command so it makes it extremely dangerous.

Think if a user had it print into chat !roulette until it got killed by the network or spam other people. VERY BAD!